Financial Data Analysis
Part 1 is to analyze financial data. The data are from the daily historical Apple stock prices(open, high, low, close and adjusted prices) from February 1, 2002 to January 31, 2017 extracted from the Yahoo Finance website. The data has logged the prices of the Apple stock everyday and comprises of the open, close, low, high and the adjusted close prices of the stock for the span of 15 years.The goal of the project is to discover an interesting trend in the apple stock prices over the past 15 years (3775 attributes) and to design and develop the best model for forecasting.
Source:     http://finance.yahoo.com/quote/AAPL/history?p=AAPL